# -V3 test fix